+++ src/cli/mod.rs
@@ -19,7 +19,7 @@ pub fn run() -> Result<(), SafeClosetError> {
 
     if let Some(path) = args.path {
         let closet = Closet::open_or_create(path)?;
-        tui::run(closet)?;
+        tui::run(closet, args.hide)?;
     } else {
         println!("Please provide as argument the path to the closet file to create or open");
     }