+++ src/tui/app.rs
@@ -10,8 +10,12 @@ use {
 /// Run the Terminal User Interface until the user decides to quit.
 ///
 /// The terminal must be already in alternate and raw mode
-pub(super) fn run(w: &mut W, closet: Closet) -> Result<(), SafeClosetError> {
-    let mut state = AppState::new(closet);
+pub(super) fn run(
+    w: &mut W,
+    closet: Closet,
+    hide_values: bool,
+) -> Result<(), SafeClosetError> {
+    let mut state = AppState::new(closet, hide_values);
     let mut view = GlobalView::default();
     view.set_area(Area::full_screen());
     view.draw(w, &mut state)?;